feat: move the geometries to Extrusion class
