+++ src/families/Base/index.ts
@@ -1,9 +1,14 @@
 import * as WEBIFC from "web-ifc";
-import * as THREE from "three";
 import { createIfcEntity, createIfcType } from "../../utils/generics";
 
-export abstract class Family {
-  protected types: any;
+type Types = {
+  REAL: (value: number) => WEBIFC.IFC4X3.IfcReal;
+  LENGTH: (value: number) => WEBIFC.IFC4X3.IfcLengthMeasure;
+  POSITIVE_LENGTH: (value: number) => WEBIFC.IFC4X3.IfcPositiveLengthMeasure;
+};
+
+export abstract class Base {
+  protected types: Types;
 
   constructor(
     protected ifcAPI: WEBIFC.IfcAPI,
@@ -140,46 +145,6 @@ export abstract class Family {
     return { placement, location };
   }
 
-  protected extrudedAreaSolid(
-    profile:
-      | WEBIFC.IFC4X3.IfcProfileDef
-      | WEBIFC.Handle<WEBIFC.IFC4X3.IfcProfileDef>,
-    position: number[] | WEBIFC.IFC4X3.IfcAxis2Placement3D,
-    direction: number[] | WEBIFC.IFC4X3.IfcDirection,
-    length: number | WEBIFC.IFC4X3.IfcPositiveLengthMeasure
-  ) {
-    if (Array.isArray(position))
-      position = this.axis2Placement3D(position).placement;
-    if (Array.isArray(direction)) direction = this.direction(direction);
-    if (typeof length === "number") length = this.positiveLength(length);
-    return createIfcEntity<typeof WEBIFC.IFC4X3.IfcExtrudedAreaSolid>(
-      this.ifcAPI,
-      this.modelID,
-      WEBIFC.IFCEXTRUDEDAREASOLID,
-      profile,
-      position,
-      direction,
-      length
-    );
-  }
-
-  protected rectangularProfile(
-    position: WEBIFC.IFC4X3.IfcAxis2Placement2D,
-    x: WEBIFC.IFC4X3.IfcPositiveLengthMeasure,
-    y: WEBIFC.IFC4X3.IfcPositiveLengthMeasure
-  ) {
-    return createIfcEntity<typeof WEBIFC.IFC4X3.IfcRectangleProfileDef>(
-      this.ifcAPI,
-      this.modelID,
-      WEBIFC.IFCRECTANGLEPROFILEDEF,
-      WEBIFC.IFC4X3.IfcProfileTypeEnum.AREA,
-      this.label("Rectangle profile"),
-      position,
-      x,
-      y
-    );
-  }
-
   protected bool(
     first: WEBIFC.IFC4X3.IfcBooleanOperand,
     second: WEBIFC.IFC4X3.IfcBooleanOperand
@@ -194,43 +159,9 @@ export abstract class Family {
     );
   }
 
-  protected vector(values: number[], type: string) {
+  protected vector(values: number[], type: keyof Types) {
     if (!this.types[type]) throw new Error(`Type not found: ${type}`);
     const action = this.types[type];
     return values.map((value) => action(value));
-  }
-
-  protected geometry(
-    data: WEBIFC.IfcGeometry
-  ): THREE.BufferGeometry<THREE.NormalBufferAttributes> {
-    const index = this.ifcAPI.GetIndexArray(
-      data.GetIndexData(),
-      data.GetIndexDataSize()
-    );
-
-    const vertexData = this.ifcAPI.GetVertexArray(
-      data.GetVertexData(),
-      data.GetVertexDataSize()
-    );
-
-    const position = new Float32Array(vertexData.length / 2);
-    const normal = new Float32Array(vertexData.length / 2);
-
-    for (let i = 0; i < vertexData.length; i += 6) {
-      position[i / 2] = vertexData[i];
-      position[i / 2 + 1] = vertexData[i + 1];
-      position[i / 2 + 2] = vertexData[i + 2];
-
-      normal[i / 2] = vertexData[i + 3];
-      normal[i / 2 + 1] = vertexData[i + 4];
-      normal[i / 2 + 2] = vertexData[i + 5];
-    }
-
-    const geometry = new THREE.BufferGeometry();
-    geometry.setAttribute("position", new THREE.BufferAttribute(position, 3));
-    geometry.setAttribute("normal", new THREE.BufferAttribute(normal, 3));
-    geometry.setIndex(Array.from(index));
-
-    return geometry;
   }
 }
